Quiz Server implementation

+++ QuizScore.java
@@ -2,9 +2,9 @@
 public class QuizScore
 {
 	public final int memberID;
-	public final double score;
+	public final int score;
 
-	public QuizScore(int memberID, double score)
+	public QuizScore(int memberID, int score)
 	{
 		this.memberID = memberID;
 		this.score = score;
